refactor: connect gemini to respective api and improve code separating concerns
